requirements.txt
# PentestLLM - keep aligned with the parent repository if you prefer a single installation
requests>=2.26.0
rich>=13.0.0
prompt_toolkit>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
